using uint8_t to free memory and speed up the process

+++ include/concertina_lib/configuration.cpp
@@ -4,16 +4,16 @@
 
 Configuration::Configuration(
   String s,
-  int o,
-  int shift,
-  int oct1,
-  int oct2,
-  int oct3,
-  int oct4,
-  int osc1,
-  int osc2,
-  int brd1,
-  int brd2
+  uint8_t o,
+  uint8_t shift,
+  uint8_t oct1,
+  uint8_t oct2,
+  uint8_t oct3,
+  uint8_t oct4,
+  uint8_t osc1,
+  uint8_t osc2,
+  uint8_t brd1,
+  uint8_t brd2
 )
 {
   presetName = s;
@@ -31,14 +31,14 @@ Configuration::Configuration(
 String Configuration::getName(){
   return presetName;
 }
-void Configuration::setOctaveOsc(int a, int b, int c, int d){
+void Configuration::setOctaveOsc(uint8_t a, uint8_t b, uint8_t c, uint8_t d){
   octaveOsc1 = a;
   octaveOsc2 = b;
   octaveBourdon1 = c;
   octaveBourdon2 = d;
 }
 
-void Configuration::setOscOct(int osc, int value){
+void Configuration::setOscOct(uint8_t osc, uint8_t value){
   if(osc == 1){
     octaveOsc1 = value;
   }else if(osc == 2){
@@ -55,22 +55,22 @@ void Configuration::setOscOct(int osc, int value){
   }
 }
 
-int Configuration::getOsc1(){return activeOsc1;}
+uint8_t Configuration::getOsc1(){return activeOsc1;}
 
-int Configuration::getOsc2(){return activeOsc2;}
+uint8_t Configuration::getOsc2(){return activeOsc2;}
 
-int Configuration::getBrd1(){return activeBrd1;}
+uint8_t Configuration::getBrd1(){return activeBrd1;}
 
-int Configuration::getBrd2(){return activeBrd2;}
+uint8_t Configuration::getBrd2(){return activeBrd2;}
 
-void Configuration::setActiveOsc(int a, int b, int c, int d){
+void Configuration::setActiveOsc(uint8_t a, uint8_t b, uint8_t c, uint8_t d){
   activeOsc1 = a;
   activeOsc2 = a;
   activeBrd1 = c;
   activeBrd2 = d;
 }
 
-void Configuration::setAllOsc(int i){
+void Configuration::setAllOsc(uint8_t i){
   activeOsc1 = i;
   activeOsc2 = i;
   activeBrd1 = i;
